feat: add http-echo app and _clusterInfo argument
http-echo is a simple local helm-chart deploying traefik/whoami and
takes information for the relevant ingress parameters.

This commit also adds a _clusterInfo object to the input of the root
app. This struct is an api that the platform (team) can use to pass
cluster-global settings that the apps can use to configure themselves
for their containing cluster. This api should be defined and developed
in a backwards compatible way by the platform team. And this api should
remain a contract between the platform and the root-apps. The root app
then has the responsibility to use it to configure it's immediate child
apps. This is demonstrated in with the 'http-echo' app