feat(util.jsonnet): Add ability to patch multi-source apps
ArgoCD Apps can have either a spec.source (legacy) field or spec.sources
(in beta as of writing, but very useful to us). So far the utils always
assumed that only spec.source was used and attempted to only patch
spec.source. If spec.sources is present, spec.source gets ignored,
therefore our changes would be lost.

This commit adds a sourceIdx argument to all relevant source modifying
util functions. makeAppPatchable automatically detects if the app is a
multiSource app and prepares internal metadata accordingly. With single
source apps, the outward behaviour will not change. For multiSource apps
it is necessary to specify sourceIdx when using util methods to
manipulate source information. Since the value of `spec.sources` is a
list, it is not possible to overlay specific source objects by simply
overlaying `spec.sources`. Instead we add a new util method
`withSourcePatchByIdx(sourcePatch, sourceIdx)`, which takes a patch object and
applies it to the source specified by `sourceIdx`. This method will
ignore sourceIdx if used on a single source app, although there is
likely no situation in which one doesn't know if an app is single or
multi-source but still knows which idx to patch if it was multi-source

Due to the way that the internal mechanism for patching multi-source
apps works, using super.spec.sources directly could yield to unexpected
results in which overrides do not apply as expected or certain overrides
by our util methods are lost entirely. To avoid this, assertions have
been added to error when this happens. To still allow patching using
outside methods and libs that target `spec.sources` directly, an
escape hatch method `withRenderedSource` is provided, which renders the
current state of the sources array back into `spec.sources` and disables
the util methods relying on this mechanism with an error.